+++ src/main/java/com/github/alfonsoleandro/mputils/files/YamlFile.java
@@ -99,7 +99,7 @@ public class YamlFile {
                 fileConfig.set(mapEntry.getKey(), mapEntry.getValue());
             }
         }
-        save();
+        save(true);
     }
 
     /**
